bigints for models

+++ src/main/scala/SMT.scala
@@ -131,8 +131,7 @@ abstract class SMTProcess(cmd : Array[String]) extends SMT {
   def getSatValue(name : String) : BigInt = {
     sendCommand("(get-value (" + name + "))")
     readLine match {
-      case numberPattern.unanchored(assignment) =>
-        assignment.toInt
+      case numberPattern.unanchored(assignment) => BigInt(assignment)
       case str => 0
     }
   }
